+++ scripts/macros.js
@@ -133,6 +133,7 @@ export {summonBeast} from './macros/2024/spells/summonBeast.js';
 export {summonCelestial, summonCelestialMace} from './macros/2024/spells/summonCelestial.js';
 export {summonFey, summonFeyFeyStep} from './macros/2024/spells/summonFey.js';
 export {summonFiend, summonFiendClaws, summonFiendDeathThroes} from './macros/2024/spells/summonFiend.js';
+export {summonUndead, summonUndeadPutrid} from './macros/2024/spells/summonUndead.js';
 export {synapticStatic} from './macros/2024/spells/synapticStatic.js';
 export {teleport} from './macros/2024/spells/teleport.js';
 export {thornWhip} from './macros/2024/spells/thornWhip.js';