+++ scripts/macros/2014/classFeatures/paladin/oathbreaker/auraOfHate.js
@@ -1,14 +1,10 @@
 import {actorUtils, combatUtils, effectUtils, genericUtils, itemUtils} from '../../../../../utils.js';
 
 async function create({trigger: {entity: item, target, identifier}}) {
+    let targetEffect = effectUtils.getEffectByIdentifier(target.actor, identifier);
+    if (targetEffect) return;
     if (itemUtils.getConfig(item, 'combatOnly') && !combatUtils.inCombat()) return;
     if (target.actor !== item.actor && !['undead', 'fiend'].includes(actorUtils.typeOrRace(target.actor))) return;
-    let targetEffect = effectUtils.getEffectByIdentifier(target.actor, identifier);
-    if (targetEffect) {
-        let effectOriginItem = await effectUtils.getOriginItem(targetEffect);
-        if (effectOriginItem.uuid === item.uuid) return;
-        await genericUtils.remove(targetEffect);
-    }
     let showIcon = itemUtils.getConfig(item, 'showIcon');
     let effectData = {
         name: item.name,
@@ -34,7 +30,6 @@ async function create({trigger: {entity: item, target, identifier}}) {
             }
         }
     };
-    // await effectUtils.createEffect(target.actor, effectData, {parentEntity: item, identifier});
     return {
         effectData,
         effectOptions: {