+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Summon Undead
 ## New Legacy Content:
 - Sentinel at Death's Door
+- Aura of Clarity
 ## Bug Fixes:
 - Shield (Modern)
 - Spirit Shield (Legacy)
@@ -18,6 +19,7 @@
 - Last Stand Armor (Legacy)
 - Summon Fey (Legacy)
 - Summon Elemental (Legacy)
+- Legacy Paldin Auras
 
 # 1.2.31 Release:
 ## Bug Fixes: