+++ CHANGELOG.md
@@ -19,7 +19,7 @@
 - Last Stand Armor (Legacy)
 - Summon Fey (Legacy)
 - Summon Elemental (Legacy)
-- Legacy Paldin Auras
+- Legacy Paladin Auras
 
 # 1.2.31 Release:
 ## Bug Fixes: